requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
sqlalchemy==2.0.36
asyncpg==0.30.0
bcrypt==4.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
